@@ -1,13 +1,11 @@
-//var React = require('react');
-import React from 'react';
-var utils = require('./utils');
+import React, {Component} from 'react';
+import utils from './utils';
 
-export default ComposedComponent => class extends React.Component {
+export default (ComposedComponent, defaultProps = {}) => class Composer extends Component {
 
     constructor(props) {
         super(props);
-        this.onChangeValidate = this.onChangeValidate.bind(this);
-        let value = this.defaultValue(this.props);
+        let value = Composer.defaultValue(this.props);
         let validationResult = utils.validate(this.props.form, value);
         this.state = {
             value: value,
@@ -16,27 +14,27 @@ export default ComposedComponent => class extends React.Component {
         };
     }
 
-    componentWillReceiveProps(nextProps) {
-      let value = this.defaultValue(nextProps);
+    static getDerivedStateFromProps(nextProps) {
+      let value = Composer.defaultValue(nextProps);
       let validationResult = utils.validate(nextProps.form, value);
-      this.setState({
-        value: value,
+      return {
+        value,
         valid: !!(validationResult.valid || !value),
         error: !validationResult.valid && value ? validationResult.error.message : null
-      });
+      }
     }
 
     /**
      * Called when <input> value changes.
-     * @param e The input element, or something.
+     * @param e The input event.
+     * @param v e.target.value equivalent when it otherwise wouldn't be there.
      */
-    onChangeValidate(e) {
-        // console.log('onChangeValidate e', e);
+    onChangeValidate = (e,v) => {
         let value = null;
         switch(this.props.form.schema.type) {
           case 'integer':
           case 'number':
-            if (e.target.value.indexOf('.') == -1) {
+            if (e.target.value.indexOf('.') === -1) {
                 value = parseInt(e.target.value);
             } else {
                 value = parseFloat(e.target.value);
@@ -49,14 +47,19 @@ export default ComposedComponent => class extends React.Component {
           case 'boolean':
             value = e.target.checked;
             break;
-          case 'object':
-          case 'date':
+          case 'tBoolean':
+            if(e.target.value !== 'yes' || e.target.value !== 'no') {
+                value = v;
+            }
+            break
           case 'array':
             value = e;
             break
+          case 'object':
           default:
             value = e.target.value;
         }
+
         //console.log('onChangeValidate this.props.form, value', this.props.form, value);
         let validationResult = utils.validate(this.props.form, value);
         this.setState({
@@ -64,19 +67,14 @@ export default ComposedComponent => class extends React.Component {
             valid: validationResult.valid,
             error: validationResult.valid ? null : validationResult.error.message
         });
-        //console.log('conhangeValidate this.props.form.key, value', this.props.form.key, value);
-        
         this.props.onChange(this.props.form.key, value);
     }
 
-    defaultValue(props) {
-        // check if there is a value in the model, if there is, display it. Otherwise, check if
-        // there is a default value, display it.
-        // console.log('Text.defaultValue key', this.props.form.key);
-        // console.log('Text.defaultValue model', this.props.model);
+    static defaultValue = (props) => {
         let value;
-        if(props.form.key)
+        if (props.form && props.form.key) {
             value = utils.selectOrSet(props.form.key, props.model);
+        }
         //console.log('Text defaultValue value = ', value);
 
         // check if there is a default value
@@ -93,11 +91,10 @@ export default ComposedComponent => class extends React.Component {
         if(!value && props.form.titleMap && props.form.titleMap[0].value) {
             value = props.form.titleMap[0].value;
         }
-        //console.log('value', value);
         return value;
     }
 
     render() {
-        return <ComposedComponent {...this.props} {...this.state} onChangeValidate={this.onChangeValidate}/>;
+        return <ComposedComponent {...defaultProps} {...this.props} {...this.state} onChangeValidate={this.onChangeValidate}/>;
     }
 };
