@@ -1,21 +1,14 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
+import React from 'react'
+import ReactDOM from 'react-dom'
+import {MuiThemeProvider, createMuiTheme} from '@material-ui/core/styles'
+import ExamplePage from './ExamplePage'
 
-var ExamplePage  = require('./ExamplePage');
+const theme = createMuiTheme({
+    typography: {
+        fontSize: 22,
+    },
+})
 
-// For me this obfuscates the location of the error when searching from console. Not sure why it's necessary.
-var warn = console.warn;
-
-console.warn = function(warning) {
-    throw new Error(warning);
-    warn.apply(console, arguments);
-};
-
-//Needed for onTouchTap
-//Can go away when react 1.0 release
-//Check this repo:
-//https://github.com/zilverline/react-tap-event-plugin
-import injectTapEventPlugin from 'react-tap-event-plugin';
-injectTapEventPlugin();
-
-ReactDOM.render(<ExamplePage />, document.getElementById("app"));
+ReactDOM.render(<MuiThemeProvider theme={theme}>
+    <ExamplePage/>
+</MuiThemeProvider>, document.getElementById('app'))
