@@ -10,61 +10,21 @@ var _react = require('react');
 
 var _react2 = _interopRequireDefault(_react);
 
+var _styles = require('@material-ui/core/styles');
+
+var _core = require('@material-ui/core');
+
 var _utils = require('./utils');
 
 var _utils2 = _interopRequireDefault(_utils);
 
-var _Number = require('./Number');
+var _cloneDeep = require('lodash/cloneDeep');
 
-var _Number2 = _interopRequireDefault(_Number);
-
-var _Text = require('./Text');
-
-var _Text2 = _interopRequireDefault(_Text);
-
-var _TextArea = require('./TextArea');
-
-var _TextArea2 = _interopRequireDefault(_TextArea);
-
-var _Select = require('./Select');
-
-var _Select2 = _interopRequireDefault(_Select);
-
-var _Radios = require('./Radios');
-
-var _Radios2 = _interopRequireDefault(_Radios);
-
-var _Date = require('./Date');
-
-var _Date2 = _interopRequireDefault(_Date);
-
-var _Checkbox = require('./Checkbox');
-
-var _Checkbox2 = _interopRequireDefault(_Checkbox);
-
-var _Help = require('./Help');
-
-var _Help2 = _interopRequireDefault(_Help);
+var _cloneDeep2 = _interopRequireDefault(_cloneDeep);
 
 var _ComposedComponent = require('./ComposedComponent');
 
 var _ComposedComponent2 = _interopRequireDefault(_ComposedComponent);
-
-var _RaisedButton = require('material-ui/RaisedButton');
-
-var _RaisedButton2 = _interopRequireDefault(_RaisedButton);
-
-var _lodash = require('lodash');
-
-var _lodash2 = _interopRequireDefault(_lodash);
-
-var _SchemaForm = require('./SchemaForm');
-
-var _SchemaForm2 = _interopRequireDefault(_SchemaForm);
-
-var _IconButton = require('material-ui/IconButton');
-
-var _IconButton2 = _interopRequireDefault(_IconButton);
 
 function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }
 
@@ -72,59 +32,60 @@ function _classCallCheck(instance, Constructor) { if (!(instance instanceof Cons
 
 function _possibleConstructorReturn(self, call) { if (!self) { throw new ReferenceError("this hasn't been initialised - super() hasn't been called"); } return call && (typeof call === "object" || typeof call === "function") ? call : self; }
 
-function _inherits(subClass, superClass) { if (typeof superClass !== "function" && superClass !== null) { throw new TypeError("Super expression must either be null or a function, not " + typeof superClass); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, enumerable: false, writable: true, configurable: true } }); if (superClass) Object.setPrototypeOf ? Object.setPrototypeOf(subClass, superClass) : subClass.__proto__ = superClass; } /**
-                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                * Created by steve on 11/09/15.
-                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                */
+function _inherits(subClass, superClass) { if (typeof superClass !== "function" && superClass !== null) { throw new TypeError("Super expression must either be null or a function, not " + typeof superClass); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, enumerable: false, writable: true, configurable: true } }); if (superClass) Object.setPrototypeOf ? Object.setPrototypeOf(subClass, superClass) : subClass.__proto__ = superClass; }
 
+var styles = function styles(theme) {
+    return {
+        arrayItem: {
+            position: 'relative',
+            border: '1px dotted #aaa',
+            padding: theme.spacing.unit,
+            marginBottom: theme.spacing.unit
+        },
+        deleteItemButton: {
+            position: 'absolute',
+            top: -theme.spacing.unit,
+            right: -theme.spacing.unit
+        },
+        addButton: {
+            marginTop: theme.spacing.unit
+        }
+    };
+};
 
-var Array = function (_React$Component) {
-    _inherits(Array, _React$Component);
+var Array = function (_Component) {
+    _inherits(Array, _Component);
+
+    _createClass(Array, null, [{
+        key: 'assignItemId',
+        value: function assignItemId(item) {
+            if (item && !item[Array.ITEM_ID]) {
+                // define hidden property with internal id
+                Object.defineProperty(item, Array.ITEM_ID, {
+                    enumerable: false,
+                    writable: true
+                });
+                item[Array.ITEM_ID] = Array._SEQUENCE++;
+            }
+            return item;
+        }
+    }]);
 
     function Array(props) {
         _classCallCheck(this, Array);
 
         var _this = _possibleConstructorReturn(this, (Array.__proto__ || Object.getPrototypeOf(Array)).call(this, props));
 
-        _this.onAppend = _this.onAppend.bind(_this);
-        _this.onDelete = _this.onDelete.bind(_this);
-        // we have the model here for the entire form, get the model for this array only
-        // and add to the state. if is empty, add an entry by calling onAppend directly.
-        _this.state = {
-            model: _utils2.default.selectOrSet(_this.props.form.key, _this.props.model) || []
-        };
-        //console.log('constructor', this.props.form.key, this.props.model, this.state.model);
-        return _this;
-    }
-
-    _createClass(Array, [{
-        key: 'componentWillReceiveProps',
-        value: function componentWillReceiveProps(nextProps) {
-            if (nextProps.model && nextProps.form && nextProps.form.key) {
-                this.setState({
-                    model: _utils2.default.selectOrSet(nextProps.form.key, nextProps.model) || []
-                });
-            }
-        }
-    }, {
-        key: 'componentDidMount',
-        value: function componentDidMount() {
-            // Always start with one empty form unless configured otherwise.
-            if (this.props.form.startEmpty !== true && this.state.model.length === 0) {
-                this.onAppend();
-            }
-        }
-    }, {
-        key: 'onAppend',
-        value: function onAppend() {
+        _this.onAppend = function () {
             //console.log('onAppend is called this.state.model', this.state.model);
-            var empty;
-            if (this.props.form && this.props.form.schema && this.props.form.schema.items) {
-                var items = this.props.form.schema.items;
+            var empty = void 0;
+            if (_this.props.form && _this.props.form.schema && _this.props.form.schema.items) {
+                var items = _this.props.form.schema.items;
                 if (items.type && items.type.indexOf('object') !== -1) {
                     empty = {};
 
                     // Check for possible defaults
-                    if (!this.props.options || this.props.options.setSchemaDefaults !== false) {
+                    if (!_this.props.options || _this.props.options.setSchemaDefaults !== false) {
                         empty = typeof items['default'] !== 'undefined' ? items['default'] : empty;
 
                         // Check for defaults further down in the schema.
@@ -140,78 +101,97 @@ var Array = function (_React$Component) {
                     }
                 } else if (items.type && items.type.indexOf('array') !== -1) {
                     empty = [];
-                    if (!this.props.options || this.props.options.setSchemaDefaults !== false) {
+                    if (!_this.props.options || _this.props.options.setSchemaDefaults !== false) {
                         empty = items['default'] || empty;
                     }
                 } else {
                     // No type? could still have defaults.
-                    if (!this.props.options || this.props.options.setSchemaDefaults !== false) {
+                    if (!_this.props.options || _this.props.options.setSchemaDefaults !== false) {
                         empty = items['default'] || empty;
                     }
                 }
             }
-            var newModel = this.state.model;
+            var newModel = _this.state.model;
+            Array.assignItemId(empty);
             newModel.push(empty);
-            this.setState({
-                model: newModel
-            });
-            this.props.onChangeValidate(this.state.model);
+            _this.setState({ model: newModel });
+            _this.props.onChangeValidate(_this.state.model);
             //console.log('After append this.state.model', newModel);
-        }
-    }, {
-        key: 'onDelete',
-        value: function onDelete(index) {
-            //console.log('onDelete is called', index);
-            var newModel = this.state.model;
-            newModel.splice(index, 1);
-            this.setState({
-                model: newModel
-            });
-            this.props.onChangeValidate(this.state.model);
-        }
-    }, {
-        key: 'setIndex',
-        value: function setIndex(index) {
+        };
+
+        _this.onDelete = function (index) {
+            return function () {
+                var newModel = _this.state.model;
+                newModel.splice(index, 1);
+                _this.setState({ model: newModel });
+                _this.props.onChangeValidate(_this.state.model);
+            };
+        };
+
+        _this.setIndex = function (index) {
             return function (form) {
                 if (form.key) {
                     form.key[form.key.indexOf('')] = index;
                 }
             };
-        }
-    }, {
-        key: 'copyWithIndex',
-        value: function copyWithIndex(form, index) {
-            var copy = _lodash2.default.cloneDeep(form);
+        };
+
+        _this.copyWithIndex = function (form, index) {
+            var copy = (0, _cloneDeep2.default)(form);
             copy.arrayIndex = index;
-            _utils2.default.traverseForm(copy, this.setIndex(index));
+            _utils2.default.traverseForm(copy, _this.setIndex(index));
             return copy;
+        };
+
+        _this.onAppend = _this.onAppend.bind(_this);
+        _this.onDelete = _this.onDelete.bind(_this);
+        // we have the model here for the entire form, get the model for this array only
+        // and add to the state. if is empty, add an entry by calling onAppend directly.
+        _this.state = {
+            model: _utils2.default.selectOrSet(_this.props.form.key, _this.props.model) || []
+            //console.log('constructor', this.props.form.key, this.props.model, this.state.model);
+        };return _this;
+    }
+
+    _createClass(Array, [{
+        key: 'componentDidMount',
+        value: function componentDidMount() {
+            // Always start with one empty form unless configured otherwise.
+            if (this.props.form.startEmpty !== true && this.state.model.length === 0) {
+                this.onAppend();
+            }
         }
     }, {
         key: 'render',
         value: function render() {
-            //console.log('Array.render', this.props.form.items, this.props.model, this.state.model);
+            var _this2 = this;
+
+            var _props = this.props,
+                classes = _props.classes,
+                form = _props.form;
+
             var arrays = [];
-            var fields = [];
             var model = this.state.model;
-            var items = this.props.form.items;
-            //console.log('fields', fields);
-            for (var i = 0; i < model.length; i++) {
-                var boundOnDelete = this.onDelete.bind(this, i);
-                var forms = this.props.form.items.map(function (form, index) {
-                    var copy = this.copyWithIndex(form, i);
-                    return this.props.builder(copy, this.props.model, index, this.props.mapper, this.props.onChange, this.props.builder);
-                }.bind(this));
-                //console.log('forms', i, forms);
+
+            var _loop = function _loop(i) {
+                var item = model[i];
+                var forms = form.items.map(function (form, index) {
+                    return _this2.props.builder(_this2.copyWithIndex(form, i), _this2.props.model, index, _this2.props.mapper, _this2.props.onChange, _this2.props.builder);
+                });
                 arrays.push(_react2.default.createElement(
-                    'li',
-                    { key: i, className: 'list-group-item' },
+                    'div',
+                    { className: classes.arrayItem, key: item && item[Array.ITEM_ID] || i },
                     _react2.default.createElement(
-                        _IconButton2.default,
-                        { iconClassName: 'material-icons', tooltip: 'Remove', onTouchTap: boundOnDelete },
-                        'clear'
+                        _core.IconButton,
+                        { onClick: _this2.onDelete(i), className: classes.deleteItemButton },
+                        _react2.default.createElement(_core.DeleteIcon, { fontSize: 'small' })
                     ),
                     forms
                 ));
+            };
+
+            for (var i = 0; i < model.length; i++) {
+                _loop(i);
             }
             return _react2.default.createElement(
                 'div',
@@ -221,36 +201,40 @@ var Array = function (_React$Component) {
                     null,
                     _react2.default.createElement(
                         'label',
-                        { className: 'control-lable' },
+                        null,
                         this.props.form.title
                     ),
                     _react2.default.createElement(
-                        'ol',
-                        { className: 'list-group' },
+                        'div',
+                        null,
                         arrays
                     )
                 ),
-                _react2.default.createElement(_RaisedButton2.default, { label: this.props.form.add || 'Add', secondary: true, onTouchTap: this.onAppend })
+                _react2.default.createElement(
+                    _core.Button,
+                    { className: classes.addButton, variant: 'contained', color: 'primary', onClick: this.onAppend },
+                    this.props.form.add || 'Add'
+                )
             );
+        }
+    }], [{
+        key: 'getDerivedStateFromProps',
+        value: function getDerivedStateFromProps(props, state) {
+            var propsKey = props.form.key;
+            if (props.form && propsKey === state.formKey && props.model && props.model[propsKey] === state.model) {
+                return null; // nothing changed
+            }
+            var model = _utils2.default.selectOrSet(propsKey, props.model) || [];
+            return {
+                formKey: propsKey,
+                model: model.map(Array.assignItemId)
+            };
         }
     }]);
 
     return Array;
-}(_react2.default.Component);
+}(_react.Component);
 
-var _default = (0, _ComposedComponent2.default)(Array);
-
-exports.default = _default;
-;
-
-var _temp = function () {
-    if (typeof __REACT_HOT_LOADER__ === 'undefined') {
-        return;
-    }
-
-    __REACT_HOT_LOADER__.register(Array, 'Array', 'src/Array.js');
-
-    __REACT_HOT_LOADER__.register(_default, 'default', 'src/Array.js');
-}();
-
-;
+Array.ITEM_ID = '_SCHEMAFORM_ITEM_ID';
+Array._SEQUENCE = 1;
+exports.default = (0, _ComposedComponent2.default)((0, _styles.withStyles)(styles)(Array));
