@@ -14,9 +14,17 @@ var _ComposedComponent = require('./ComposedComponent');
 
 var _ComposedComponent2 = _interopRequireDefault(_ComposedComponent);
 
-var _Checkbox = require('material-ui/Checkbox');
+var _Checkbox = require('@material-ui/core/Checkbox');
 
 var _Checkbox2 = _interopRequireDefault(_Checkbox);
+
+var _FormGroup = require('@material-ui/core/FormGroup');
+
+var _FormGroup2 = _interopRequireDefault(_FormGroup);
+
+var _FormControlLabel = require('@material-ui/core/FormControlLabel');
+
+var _FormControlLabel2 = _interopRequireDefault(_FormControlLabel);
 
 function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }
 
@@ -24,31 +32,20 @@ function _classCallCheck(instance, Constructor) { if (!(instance instanceof Cons
 
 function _possibleConstructorReturn(self, call) { if (!self) { throw new ReferenceError("this hasn't been initialised - super() hasn't been called"); } return call && (typeof call === "object" || typeof call === "function") ? call : self; }
 
-function _inherits(subClass, superClass) { if (typeof superClass !== "function" && superClass !== null) { throw new TypeError("Super expression must either be null or a function, not " + typeof superClass); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, enumerable: false, writable: true, configurable: true } }); if (superClass) Object.setPrototypeOf ? Object.setPrototypeOf(subClass, superClass) : subClass.__proto__ = superClass; } /**
-                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                * Created by steve on 20/09/15.
-                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                */
+function _inherits(subClass, superClass) { if (typeof superClass !== "function" && superClass !== null) { throw new TypeError("Super expression must either be null or a function, not " + typeof superClass); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, enumerable: false, writable: true, configurable: true } }); if (superClass) Object.setPrototypeOf ? Object.setPrototypeOf(subClass, superClass) : subClass.__proto__ = superClass; }
 
+var FormCheckbox = function (_Component) {
+    _inherits(FormCheckbox, _Component);
 
-var styles = {
-    error: {
-        color: "rgb(244, 67, 54)",
-        fontSize: "12px",
-        lineHeight: "12px"
-    },
-    checkbox: {
-        marginTop: "14px",
-        marginBottom: "0px",
-        height: "72px"
-    }
-};
+    function FormCheckbox(props) {
+        _classCallCheck(this, FormCheckbox);
 
-var Checkbox2 = function (_React$Component) {
-    _inherits(Checkbox2, _React$Component);
+        var _this = _possibleConstructorReturn(this, (FormCheckbox.__proto__ || Object.getPrototypeOf(FormCheckbox)).call(this, props));
 
-    function Checkbox2(props) {
-        _classCallCheck(this, Checkbox2);
-
-        var _this = _possibleConstructorReturn(this, (Checkbox2.__proto__ || Object.getPrototypeOf(Checkbox2)).call(this, props));
+        _this.handleChange = function (e) {
+            _this.props.onChangeValidate(e, 'yes');
+            // this.props.onChangeValidate(e, !this.props.value)
+        };
 
         var _this$props = _this.props,
             model = _this$props.model,
@@ -56,56 +53,33 @@ var Checkbox2 = function (_React$Component) {
         var key = form.key;
         //If a boolean is stored, use it; if not, if a boolean is defined as schema's default, use it.
 
-        var value = _.isBoolean(_this.props.value) ? _this.props.value : _.isBoolean(form.schema.default) ? form.schema.default : undefined;
+        var value = typeof _this.props.value === 'boolean' ? _this.props.value : typeof form.schema.default === 'boolean' ? form.schema.default : undefined;
         _this.props.setDefault(key, model, form, value);
         return _this;
     }
 
-    _createClass(Checkbox2, [{
+    _createClass(FormCheckbox, [{
         key: 'render',
         value: function render() {
-            var _this2 = this;
-
             return _react2.default.createElement(
-                'div',
-                { className: this.props.form.className, style: styles.checkbox },
-                _react2.default.createElement(_Checkbox2.default, {
-                    name: this.props.form.key.slice(-1)[0],
-                    value: this.props.form.key.slice(-1)[0],
-                    checked: this.props.value,
+                _FormGroup2.default,
+                { row: true },
+                _react2.default.createElement(_FormControlLabel2.default, {
+                    className: this.props.form.className,
                     label: this.props.form.title,
-                    disabled: this.props.form.readonly,
-                    onCheck: function onCheck(e, checked) {
-                        _this2.props.onChangeValidate(e);
-                    }
-                }),
-                _react2.default.createElement(
-                    'span',
-                    { style: styles.error },
-                    this.props.errorText
-                )
+                    control: _react2.default.createElement(_Checkbox2.default, {
+                        name: this.props.form.key.slice(-1)[0],
+                        value: this.props.form.key.slice(-1)[0],
+                        checked: this.props.value || false,
+                        disabled: this.props.form.readonly,
+                        onChange: this.handleChange
+                    })
+                })
             );
         }
     }]);
 
-    return Checkbox2;
-}(_react2.default.Component);
+    return FormCheckbox;
+}(_react.Component);
 
-var _default = (0, _ComposedComponent2.default)(Checkbox2);
-
-exports.default = _default;
-;
-
-var _temp = function () {
-    if (typeof __REACT_HOT_LOADER__ === 'undefined') {
-        return;
-    }
-
-    __REACT_HOT_LOADER__.register(styles, 'styles', 'src/Checkbox.js');
-
-    __REACT_HOT_LOADER__.register(Checkbox2, 'Checkbox2', 'src/Checkbox.js');
-
-    __REACT_HOT_LOADER__.register(_default, 'default', 'src/Checkbox.js');
-}();
-
-;
+exports.default = (0, _ComposedComponent2.default)(FormCheckbox);
