@@ -10,49 +10,11 @@ var _react = require('react');
 
 var _react2 = _interopRequireDefault(_react);
 
-var _utils = require('./utils');
+var _FormControl = require('@material-ui/core/FormControl');
 
-var _utils2 = _interopRequireDefault(_utils);
+var _FormControl2 = _interopRequireDefault(_FormControl);
 
-var _Number = require('./Number');
-
-var _Number2 = _interopRequireDefault(_Number);
-
-var _Text = require('./Text');
-
-var _Text2 = _interopRequireDefault(_Text);
-
-var _TextArea = require('./TextArea');
-
-var _TextArea2 = _interopRequireDefault(_TextArea);
-
-var _Select = require('./Select');
-
-var _Select2 = _interopRequireDefault(_Select);
-
-var _Radios = require('./Radios');
-
-var _Radios2 = _interopRequireDefault(_Radios);
-
-var _Date = require('./Date');
-
-var _Date2 = _interopRequireDefault(_Date);
-
-var _Checkbox = require('./Checkbox');
-
-var _Checkbox2 = _interopRequireDefault(_Checkbox);
-
-var _Help = require('./Help');
-
-var _Help2 = _interopRequireDefault(_Help);
-
-var _RaisedButton = require('material-ui/RaisedButton');
-
-var _RaisedButton2 = _interopRequireDefault(_RaisedButton);
-
-var _lodash = require('lodash');
-
-var _lodash2 = _interopRequireDefault(_lodash);
+var _styles = require('@material-ui/core/styles');
 
 function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }
 
@@ -60,13 +22,18 @@ function _classCallCheck(instance, Constructor) { if (!(instance instanceof Cons
 
 function _possibleConstructorReturn(self, call) { if (!self) { throw new ReferenceError("this hasn't been initialised - super() hasn't been called"); } return call && (typeof call === "object" || typeof call === "function") ? call : self; }
 
-function _inherits(subClass, superClass) { if (typeof superClass !== "function" && superClass !== null) { throw new TypeError("Super expression must either be null or a function, not " + typeof superClass); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, enumerable: false, writable: true, configurable: true } }); if (superClass) Object.setPrototypeOf ? Object.setPrototypeOf(subClass, superClass) : subClass.__proto__ = superClass; } /**
-                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                * Created by steve on 11/09/15.
-                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                */
+function _inherits(subClass, superClass) { if (typeof superClass !== "function" && superClass !== null) { throw new TypeError("Super expression must either be null or a function, not " + typeof superClass); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, enumerable: false, writable: true, configurable: true } }); if (superClass) Object.setPrototypeOf ? Object.setPrototypeOf(subClass, superClass) : subClass.__proto__ = superClass; }
 
+var styles = function styles(theme) {
+    return {
+        root: {
+            marginTop: theme.spacing.unit
+        }
+    };
+};
 
-var FieldSet = function (_React$Component) {
-    _inherits(FieldSet, _React$Component);
+var FieldSet = function (_Component) {
+    _inherits(FieldSet, _Component);
 
     function FieldSet() {
         _classCallCheck(this, FieldSet);
@@ -77,20 +44,22 @@ var FieldSet = function (_React$Component) {
     _createClass(FieldSet, [{
         key: 'render',
         value: function render() {
-            //console.log('FieldSet.render', this.props);
+            var _props = this.props,
+                form = _props.form,
+                mapper = _props.mapper,
+                builder = _props.builder,
+                model = _props.model,
+                onChange = _props.onChange,
+                classes = _props.classes;
             // now render all the items in the fieldset
-            var forms = this.props.form.items.map(function (form, index) {
-                return this.props.builder(form, this.props.model, index, this.props.mapper, this.props.onChange, this.props.builder);
-            }.bind(this));
+
+            var forms = form.items.map(function (f, index) {
+                return builder(f, model, index, onChange, mapper, builder);
+            });
 
             return _react2.default.createElement(
-                'fieldset',
-                null,
-                _react2.default.createElement(
-                    'legend',
-                    null,
-                    this.props.form.title
-                ),
+                _FormControl2.default,
+                { component: 'fieldset', className: classes.root },
                 _react2.default.createElement(
                     'div',
                     null,
@@ -101,20 +70,6 @@ var FieldSet = function (_React$Component) {
     }]);
 
     return FieldSet;
-}(_react2.default.Component);
+}(_react.Component);
 
-var _default = FieldSet;
-exports.default = _default;
-;
-
-var _temp = function () {
-    if (typeof __REACT_HOT_LOADER__ === 'undefined') {
-        return;
-    }
-
-    __REACT_HOT_LOADER__.register(FieldSet, 'FieldSet', 'src/FieldSet.js');
-
-    __REACT_HOT_LOADER__.register(_default, 'default', 'src/FieldSet.js');
-}();
-
-;
+exports.default = (0, _styles.withStyles)(styles)(FieldSet);
