@@ -14,9 +14,7 @@ var _ComposedComponent = require('./ComposedComponent');
 
 var _ComposedComponent2 = _interopRequireDefault(_ComposedComponent);
 
-var _TextField = require('material-ui/TextField');
-
-var _TextField2 = _interopRequireDefault(_TextField);
+var _core = require('@material-ui/core');
 
 function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }
 
@@ -24,46 +22,76 @@ function _classCallCheck(instance, Constructor) { if (!(instance instanceof Cons
 
 function _possibleConstructorReturn(self, call) { if (!self) { throw new ReferenceError("this hasn't been initialised - super() hasn't been called"); } return call && (typeof call === "object" || typeof call === "function") ? call : self; }
 
-function _inherits(subClass, superClass) { if (typeof superClass !== "function" && superClass !== null) { throw new TypeError("Super expression must either be null or a function, not " + typeof superClass); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, enumerable: false, writable: true, configurable: true } }); if (superClass) Object.setPrototypeOf ? Object.setPrototypeOf(subClass, superClass) : subClass.__proto__ = superClass; } /**
-                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                * Created by steve on 15/09/15.
-                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                */
-
+function _inherits(subClass, superClass) { if (typeof superClass !== "function" && superClass !== null) { throw new TypeError("Super expression must either be null or a function, not " + typeof superClass); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, enumerable: false, writable: true, configurable: true } }); if (superClass) Object.setPrototypeOf ? Object.setPrototypeOf(subClass, superClass) : subClass.__proto__ = superClass; }
 
 /**
  * There is no default number picker as part of Material-UI.
  * Instead, use a TextField and validate.
  */
-var Number = function (_React$Component) {
-    _inherits(Number, _React$Component);
+var Number = function (_Component) {
+    _inherits(Number, _Component);
 
     function Number(props) {
         _classCallCheck(this, Number);
 
         var _this = _possibleConstructorReturn(this, (Number.__proto__ || Object.getPrototypeOf(Number)).call(this, props));
 
-        _this.preValidationCheck = _this.preValidationCheck.bind(_this);
+        _this.isNumeric = function (n) {
+            return !isNaN(parseFloat(n)) && isFinite(n);
+        };
+
+        _this.isEmpty = function (n) {
+            return !n || 0 === n.length;
+        };
+
+        _this.preValidationCheck = function (e) {
+            if (_this.isNumeric(e.target.value)) {
+                _this.setState({
+                    lastSuccessfulValue: e.target.value
+                });
+                _this.props.onChangeValidate(e);
+            } else if (_this.isEmpty(e.target.value)) {
+                _this.setState({
+                    lastSuccessfulValue: e.target.value
+                });
+                _this.props.onChangeValidate(e);
+            } else {
+                _this.numberField.current.value = _this.state.lastSuccessfulValue;
+            }
+        };
+
         _this.state = {
             lastSuccessfulValue: _this.props.value
         };
+        _this.numberField = _react2.default.createRef();
         return _this;
     }
 
     _createClass(Number, [{
-        key: 'componentWillReceiveProps',
-        value: function componentWillReceiveProps(nextProps) {
-            this.setState({
+        key: 'render',
+        value: function render() {
+            var _props = this.props,
+                form = _props.form,
+                error = _props.error;
+
+            return _react2.default.createElement(_core.TextField, {
+                type: form.type,
+                label: form.title,
+                placeholder: form.placeholder,
+                helperText: error || form.description,
+                error: !!error,
+                onChange: this.preValidationCheck,
+                value: this.state.lastSuccessfulValue,
+                ref: this.numberField,
+                disabled: form.readonly,
+                fullWidth: true });
+        }
+    }], [{
+        key: 'getDerivedStateFromProps',
+        value: function getDerivedStateFromProps(nextProps) {
+            return {
                 lastSuccessfulValue: nextProps.value
-            });
-        }
-    }, {
-        key: 'isNumeric',
-        value: function isNumeric(n) {
-            return !isNaN(parseFloat(n)) && isFinite(n);
-        }
-    }, {
-        key: 'isEmpty',
-        value: function isEmpty(n) {
-            return !n || 0 === n.length;
+            };
         }
 
         /**
@@ -71,59 +99,9 @@ var Number = function (_React$Component) {
          * @param e
          */
 
-    }, {
-        key: 'preValidationCheck',
-        value: function preValidationCheck(e) {
-            if (this.isNumeric(e.target.value)) {
-                this.setState({
-                    lastSuccessfulValue: e.target.value
-                });
-                this.props.onChangeValidate(e);
-            } else if (this.isEmpty(e.target.value)) {
-                this.setState({
-                    lastSuccessfulValue: e.target.value
-                });
-                this.props.onChangeValidate(e);
-            } else {
-                this.refs.numberField.value = this.state.lastSuccessfulValue;
-            }
-        }
-    }, {
-        key: 'render',
-        value: function render() {
-            return _react2.default.createElement(
-                'div',
-                { className: this.props.form.htmlClass },
-                _react2.default.createElement(_TextField2.default, {
-                    type: this.props.form.type,
-                    floatingLabelText: this.props.form.title,
-                    hintText: this.props.form.placeholder,
-                    errorText: this.props.error || this.props.errorText,
-                    onChange: this.preValidationCheck,
-                    value: this.state.lastSuccessfulValue,
-                    ref: 'numberField',
-                    disabled: this.props.form.readonly,
-                    style: this.props.form.style || { width: '100%' } })
-            );
-        }
     }]);
 
     return Number;
-}(_react2.default.Component);
+}(_react.Component);
 
-var _default = (0, _ComposedComponent2.default)(Number);
-
-exports.default = _default;
-;
-
-var _temp = function () {
-    if (typeof __REACT_HOT_LOADER__ === 'undefined') {
-        return;
-    }
-
-    __REACT_HOT_LOADER__.register(Number, 'Number', 'src/Number.js');
-
-    __REACT_HOT_LOADER__.register(_default, 'default', 'src/Number.js');
-}();
-
-;
+exports.default = (0, _ComposedComponent2.default)(Number);
