@@ -10,25 +10,13 @@ var _react = require('react');
 
 var _react2 = _interopRequireDefault(_react);
 
-var _utils = require('./utils');
+var _core = require('@material-ui/core');
 
-var _utils2 = _interopRequireDefault(_utils);
-
-var _classnames = require('classnames');
-
-var _classnames2 = _interopRequireDefault(_classnames);
+var _styles = require('@material-ui/core/styles');
 
 var _ComposedComponent = require('./ComposedComponent');
 
 var _ComposedComponent2 = _interopRequireDefault(_ComposedComponent);
-
-var _RadioButton = require('material-ui/RadioButton');
-
-var _RadioButton2 = _interopRequireDefault(_RadioButton);
-
-var _RadioButtonGroup = require('material-ui/RadioButton/RadioButtonGroup');
-
-var _RadioButtonGroup2 = _interopRequireDefault(_RadioButtonGroup);
 
 function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }
 
@@ -38,8 +26,19 @@ function _possibleConstructorReturn(self, call) { if (!self) { throw new Referen
 
 function _inherits(subClass, superClass) { if (typeof superClass !== "function" && superClass !== null) { throw new TypeError("Super expression must either be null or a function, not " + typeof superClass); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, enumerable: false, writable: true, configurable: true } }); if (superClass) Object.setPrototypeOf ? Object.setPrototypeOf(subClass, superClass) : subClass.__proto__ = superClass; }
 
-var Radios = function (_React$Component) {
-    _inherits(Radios, _React$Component);
+var styles = function styles(theme) {
+    return {
+        formControl: {
+            marginTop: theme.spacing.unit
+        },
+        group: {
+            margin: theme.spacing.unit + 'px 0'
+        }
+    };
+};
+
+var Radios = function (_Component) {
+    _inherits(Radios, _Component);
 
     function Radios() {
         _classCallCheck(this, Radios);
@@ -50,47 +49,39 @@ var Radios = function (_React$Component) {
     _createClass(Radios, [{
         key: 'render',
         value: function render() {
-            var items = this.props.form.titleMap.map(function (item, index) {
-                return _react2.default.createElement(_RadioButton2.default, { label: item.name,
-                    value: item.value,
-                    key: index,
-                    disabled: this.props.form.readonly
-                });
-            }.bind(this));
+            var _this2 = this;
+
+            var classes = this.props.classes;
 
             return _react2.default.createElement(
-                'span',
-                { className: this.props.form.htmlClass },
+                _core.FormControl,
+                { component: 'fieldset', className: classes.formControl },
                 _react2.default.createElement(
-                    'label',
-                    { className: 'control-lable' },
+                    _core.FormLabel,
+                    { component: 'legend' },
                     this.props.form.title
                 ),
                 _react2.default.createElement(
-                    _RadioButtonGroup2.default,
-                    { defaultSelected: this.props.value, name: this.props.form.title, onChange: this.props.onChangeValidate },
-                    items
+                    _core.RadioGroup,
+                    {
+                        value: this.props.value,
+                        name: this.props.form.title,
+                        onChange: this.props.onChangeValidate,
+                        className: classes.group },
+                    this.props.form.titleMap.map(function (item, index) {
+                        return _react2.default.createElement(_core.FormControlLabel, {
+                            key: index,
+                            label: item.name,
+                            value: item.value,
+                            disabled: _this2.props.form.readonly,
+                            control: _react2.default.createElement(_core.Radio, { checked: _this2.props.value === item.value }) });
+                    })
                 )
             );
         }
     }]);
 
     return Radios;
-}(_react2.default.Component);
+}(_react.Component);
 
-var _default = (0, _ComposedComponent2.default)(Radios);
-
-exports.default = _default;
-;
-
-var _temp = function () {
-    if (typeof __REACT_HOT_LOADER__ === 'undefined') {
-        return;
-    }
-
-    __REACT_HOT_LOADER__.register(Radios, 'Radios', 'src/Radios.js');
-
-    __REACT_HOT_LOADER__.register(_default, 'default', 'src/Radios.js');
-}();
-
-;
+exports.default = (0, _ComposedComponent2.default)((0, _styles.withStyles)(styles)(Radios));
