@@ -14,17 +14,9 @@ var _ComposedComponent = require('./ComposedComponent');
 
 var _ComposedComponent2 = _interopRequireDefault(_ComposedComponent);
 
-var _MenuItem = require('material-ui/MenuItem');
-
-var _MenuItem2 = _interopRequireDefault(_MenuItem);
-
-var _SelectField = require('material-ui/SelectField');
-
-var _SelectField2 = _interopRequireDefault(_SelectField);
+var _core = require('@material-ui/core');
 
 var _lodash = require('lodash');
-
-var _lodash2 = _interopRequireDefault(_lodash);
 
 function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }
 
@@ -32,20 +24,17 @@ function _classCallCheck(instance, Constructor) { if (!(instance instanceof Cons
 
 function _possibleConstructorReturn(self, call) { if (!self) { throw new ReferenceError("this hasn't been initialised - super() hasn't been called"); } return call && (typeof call === "object" || typeof call === "function") ? call : self; }
 
-function _inherits(subClass, superClass) { if (typeof superClass !== "function" && superClass !== null) { throw new TypeError("Super expression must either be null or a function, not " + typeof superClass); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, enumerable: false, writable: true, configurable: true } }); if (superClass) Object.setPrototypeOf ? Object.setPrototypeOf(subClass, superClass) : subClass.__proto__ = superClass; } /**
-                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                * Created by steve on 15/09/15.
-                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                */
+function _inherits(subClass, superClass) { if (typeof superClass !== "function" && superClass !== null) { throw new TypeError("Super expression must either be null or a function, not " + typeof superClass); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, enumerable: false, writable: true, configurable: true } }); if (superClass) Object.setPrototypeOf ? Object.setPrototypeOf(subClass, superClass) : subClass.__proto__ = superClass; }
 
-
-var Select = function (_React$Component) {
-    _inherits(Select, _React$Component);
+var Select = function (_Component) {
+    _inherits(Select, _Component);
 
     function Select(props) {
         _classCallCheck(this, Select);
 
         var _this = _possibleConstructorReturn(this, (Select.__proto__ || Object.getPrototypeOf(Select)).call(this, props));
 
-        _this.onSelected = _this.onSelected.bind(_this);
+        _initialiseProps.call(_this);
 
         var _this$props = _this.props,
             model = _this$props.model,
@@ -53,9 +42,9 @@ var Select = function (_React$Component) {
         var key = form.key;
 
 
-        var storedValue = model && _this.getModelKey(model, key) || false;
+        var storedValue = model && Select.getModelKey(model, key) || false;
         var defaultValue = form.schema.default || false;
-        var value = !_lodash2.default.isEmpty(storedValue) && storedValue || defaultValue;
+        var value = !(0, _lodash.isEmpty)(storedValue) && storedValue || defaultValue;
 
         _this.props.setDefault(key, model, form, value);
         _this.state = {
@@ -65,81 +54,74 @@ var Select = function (_React$Component) {
     }
 
     _createClass(Select, [{
-        key: 'componentWillReceiveProps',
-        value: function componentWillReceiveProps(nextProps) {
-            if (nextProps.model && nextProps.form.key) {
-                this.setState({
-                    currentValue: this.getModelKey(nextProps.model, nextProps.form.key) || (nextProps.form.titleMap != null ? nextProps.form.titleMap[0].value : '')
-                });
-            }
-        }
-    }, {
-        key: 'getModelKey',
-        value: function getModelKey(model, key) {
-            if (Array.isArray(key)) {
-                return key.reduce(function (cur, nxt) {
-                    return cur[nxt] || {};
-                }, model);
-            } else {
-                return model[key];
-            }
-        }
-    }, {
-        key: 'onSelected',
-        value: function onSelected(event, selectedIndex, menuItem) {
-
-            this.setState({
-                currentValue: menuItem
-            });
-            event.target.value = menuItem;
-            this.props.onChangeValidate(event);
-        }
-    }, {
         key: 'render',
         value: function render() {
-            var menuItems = this.props.form.titleMap.map(function (item, idx) {
-                return _react2.default.createElement(_MenuItem2.default, { key: idx,
-                    primaryText: item.name,
-                    value: item.value });
-            });
+            var form = this.props.form;
 
+            var menuItems = form.titleMap.map(function (item, idx) {
+                return _react2.default.createElement(
+                    _core.MenuItem,
+                    { key: idx, value: item.value },
+                    item.name
+                );
+            });
             return _react2.default.createElement(
-                'div',
-                { className: this.props.form.htmlClass },
+                _core.FormControl,
+                { style: { width: '100%' } },
                 _react2.default.createElement(
-                    _SelectField2.default,
+                    _core.InputLabel,
+                    null,
+                    form.title
+                ),
+                _react2.default.createElement(
+                    _core.MuiSelect,
                     {
-                        value: this.state.currentValue,
-                        floatingLabelText: this.props.form.title,
-                        disabled: this.props.form.readonly,
-                        onChange: this.onSelected,
-                        fullWidth: true },
+                        disabled: form.readonly,
+                        value: this.state.currentValue || '',
+                        onChange: this.onSelected },
                     menuItems
                 )
             );
         }
+    }], [{
+        key: 'getDerivedStateFromProps',
+        value: function getDerivedStateFromProps(props) {
+            if (props.model && props.form.key) {
+                return {
+                    currentValue: Select.getModelValue(props.model, props.form)
+                };
+            }
+        }
     }]);
 
     return Select;
-}(_react2.default.Component);
+}(_react.Component);
 
-// Select.propTypes = {
-//
-// };
-
-var _default = (0, _ComposedComponent2.default)(Select);
-
-exports.default = _default;
-;
-
-var _temp = function () {
-    if (typeof __REACT_HOT_LOADER__ === 'undefined') {
-        return;
+Select.getModelKey = function (model, key) {
+    if (Array.isArray(key)) {
+        return key.reduce(function (cur, nxt) {
+            return cur[nxt] || {};
+        }, model);
+    } else {
+        return model[key];
     }
+};
 
-    __REACT_HOT_LOADER__.register(Select, 'Select', 'src/Select.js');
+Select.getModelValue = function (model, _ref) {
+    var key = _ref.key,
+        titleMap = _ref.titleMap;
+    return Select.getModelKey(model, key) || (titleMap != null ? titleMap[0].value : '');
+};
 
-    __REACT_HOT_LOADER__.register(_default, 'default', 'src/Select.js');
-}();
+var _initialiseProps = function _initialiseProps() {
+    var _this2 = this;
 
-;
+    this.onSelected = function (_ref2) {
+        var currentValue = _ref2.target.value;
+
+        _this2.setState({ currentValue: currentValue });
+        _this2.props.onChangeValidate(event);
+    };
+};
+
+exports.default = (0, _ComposedComponent2.default)(Select);
