@@ -14,9 +14,7 @@ var _ComposedComponent = require('./ComposedComponent');
 
 var _ComposedComponent2 = _interopRequireDefault(_ComposedComponent);
 
-var _TextField = require('material-ui/TextField');
-
-var _TextField2 = _interopRequireDefault(_TextField);
+var _core = require('@material-ui/core');
 
 function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }
 
@@ -24,13 +22,10 @@ function _classCallCheck(instance, Constructor) { if (!(instance instanceof Cons
 
 function _possibleConstructorReturn(self, call) { if (!self) { throw new ReferenceError("this hasn't been initialised - super() hasn't been called"); } return call && (typeof call === "object" || typeof call === "function") ? call : self; }
 
-function _inherits(subClass, superClass) { if (typeof superClass !== "function" && superClass !== null) { throw new TypeError("Super expression must either be null or a function, not " + typeof superClass); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, enumerable: false, writable: true, configurable: true } }); if (superClass) Object.setPrototypeOf ? Object.setPrototypeOf(subClass, superClass) : subClass.__proto__ = superClass; } /**
-                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                * Created by steve on 15/09/15.
-                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                */
+function _inherits(subClass, superClass) { if (typeof superClass !== "function" && superClass !== null) { throw new TypeError("Super expression must either be null or a function, not " + typeof superClass); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, enumerable: false, writable: true, configurable: true } }); if (superClass) Object.setPrototypeOf ? Object.setPrototypeOf(subClass, superClass) : subClass.__proto__ = superClass; }
 
-
-var Text = function (_React$Component) {
-    _inherits(Text, _React$Component);
+var Text = function (_Component) {
+    _inherits(Text, _Component);
 
     function Text(props) {
         _classCallCheck(this, Text);
@@ -51,39 +46,27 @@ var Text = function (_React$Component) {
     _createClass(Text, [{
         key: 'render',
         value: function render() {
-            //console.log('Text props', this.props);
-            return _react2.default.createElement(
-                'div',
-                { className: this.props.form.htmlClass },
-                _react2.default.createElement(_TextField2.default, {
-                    type: this.props.form.type,
-                    floatingLabelText: this.props.form.title,
-                    hintText: this.props.form.placeholder,
-                    errorText: this.props.error || this.props.errorText,
-                    onChange: this.props.onChangeValidate,
-                    defaultValue: this.props.value,
-                    disabled: this.props.form.readonly,
-                    style: this.props.form.style || { width: '100%' } })
-            );
+            var _props = this.props,
+                form = _props.form,
+                value = _props.value,
+                error = _props.error,
+                onChangeValidate = _props.onChangeValidate;
+
+            return _react2.default.createElement(_core.TextField, {
+                type: form.type,
+                label: form.title,
+                placeholder: form.placeholder,
+                helperText: error || form.description,
+                error: !!error,
+                onChange: onChangeValidate,
+                value: value,
+                disabled: form.readonly,
+                fullWidth: true
+            });
         }
     }]);
 
     return Text;
-}(_react2.default.Component);
+}(_react.Component);
 
-var _default = (0, _ComposedComponent2.default)(Text);
-
-exports.default = _default;
-;
-
-var _temp = function () {
-    if (typeof __REACT_HOT_LOADER__ === 'undefined') {
-        return;
-    }
-
-    __REACT_HOT_LOADER__.register(Text, 'Text', 'src/Text.js');
-
-    __REACT_HOT_LOADER__.register(_default, 'default', 'src/Text.js');
-}();
-
-;
+exports.default = (0, _ComposedComponent2.default)(Text);
