@@ -14,9 +14,9 @@ var _ComposedComponent = require('./ComposedComponent');
 
 var _ComposedComponent2 = _interopRequireDefault(_ComposedComponent);
 
-var _TextField = require('material-ui/TextField');
+var _core = require('@material-ui/core');
 
-var _TextField2 = _interopRequireDefault(_TextField);
+var _utils = require('./utils');
 
 function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }
 
@@ -24,10 +24,7 @@ function _classCallCheck(instance, Constructor) { if (!(instance instanceof Cons
 
 function _possibleConstructorReturn(self, call) { if (!self) { throw new ReferenceError("this hasn't been initialised - super() hasn't been called"); } return call && (typeof call === "object" || typeof call === "function") ? call : self; }
 
-function _inherits(subClass, superClass) { if (typeof superClass !== "function" && superClass !== null) { throw new TypeError("Super expression must either be null or a function, not " + typeof superClass); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, enumerable: false, writable: true, configurable: true } }); if (superClass) Object.setPrototypeOf ? Object.setPrototypeOf(subClass, superClass) : subClass.__proto__ = superClass; } /**
-                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                * Created by steve on 15/09/15.
-                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                */
-
+function _inherits(subClass, superClass) { if (typeof superClass !== "function" && superClass !== null) { throw new TypeError("Super expression must either be null or a function, not " + typeof superClass); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, enumerable: false, writable: true, configurable: true } }); if (superClass) Object.setPrototypeOf ? Object.setPrototypeOf(subClass, superClass) : subClass.__proto__ = superClass; }
 
 var TextArea = function (_React$Component) {
     _inherits(TextArea, _React$Component);
@@ -41,44 +38,31 @@ var TextArea = function (_React$Component) {
     _createClass(TextArea, [{
         key: 'render',
         value: function render() {
-            // FIXME: Obviously fix rowsMax eventually..
-            //console.log('TextArea', this.props.form);
-            return _react2.default.createElement(
-                'div',
-                { className: this.props.form.htmlClass },
-                _react2.default.createElement(_TextField2.default, {
-                    type: this.props.form.type,
-                    floatingLabelText: this.props.form.title,
-                    hintText: this.props.form.placeholder,
-                    onChange: this.props.onChangeValidate,
-                    errorText: this.props.error || this.props.errorText,
-                    defaultValue: this.props.value,
-                    multiLine: true,
-                    rows: this.props.form.rows,
-                    rowsMax: this.props.form.rowsMax,
-                    disabled: this.props.form.readonly,
-                    style: this.props.form.style || { width: '100%' }
-                })
-            );
+            var _props = this.props,
+                form = _props.form,
+                value = _props.value,
+                error = _props.error,
+                onChangeValidate = _props.onChangeValidate;
+
+            value = (0, _utils.selectOrSet)(this.props.form.key, this.props.model) || '';
+            return _react2.default.createElement(_core.TextField, {
+                type: form.type,
+                label: form.title,
+                placeholder: form.placeholder,
+                helperText: error || form.description,
+                onChange: onChangeValidate,
+                error: !!error,
+                value: value,
+                multiline: true,
+                rows: form.rows,
+                rowsMax: form.rowsMax,
+                disabled: form.readonly,
+                fullWidth: true
+            });
         }
     }]);
 
     return TextArea;
 }(_react2.default.Component);
 
-var _default = (0, _ComposedComponent2.default)(TextArea);
-
-exports.default = _default;
-;
-
-var _temp = function () {
-    if (typeof __REACT_HOT_LOADER__ === 'undefined') {
-        return;
-    }
-
-    __REACT_HOT_LOADER__.register(TextArea, 'TextArea', 'src/TextArea.js');
-
-    __REACT_HOT_LOADER__.register(_default, 'default', 'src/TextArea.js');
-}();
-
-;
+exports.default = (0, _ComposedComponent2.default)(TextArea);
