@@ -1,43 +1,36 @@
 'use strict';
 
-jest.dontMock('../SchemaForm');
-jest.dontMock('../utils');
-jest.dontMock('lodash');
+var _react = require('react');
 
-// var React = require('react');
-var TestUtils = require('react-addons-test-utils');
-var SchemaForm = require('../SchemaForm');
+var _react2 = _interopRequireDefault(_react);
 
-describe('SchemaForm', function () {
+var _SchemaForm = require('../SchemaForm');
 
-  beforeEach(function () {});
+var _SchemaForm2 = _interopRequireDefault(_SchemaForm);
 
-  it('shows SchemaForm', function () {
-    var shallowRenderer = TestUtils.createRenderer();
-    var cfg = {
-      form: {},
-      schema: {
-        'type': 'object'
-      },
-      model: {},
-      mapper: {}
-    };
-    shallowRenderer.render(React.createElement(SchemaForm, {
-      schema: cfg.schema,
-      mapper: cfg.mapper
-    }));
+var _shallow = require('react-test-renderer/shallow');
 
-    var result = shallowRenderer.getRenderOutput();
-    console.log('result = ', result.props);
-    expect(result.props.children).toEqual('SchemaForm');
-  });
+var _shallow2 = _interopRequireDefault(_shallow);
+
+function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }
+
+describe('SchemaForm test', function () {
+
+    it('shows SchemaForm', function () {
+        var cfg = {
+            form: {},
+            schema: {
+                'type': 'object'
+            },
+            model: {},
+            mapper: {}
+        };
+        _shallow2.default.render(_react2.default.createElement(_SchemaForm2.default, {
+            schema: cfg.schema,
+            mapper: cfg.mapper
+        }));
+        var result = _shallow2.default.getRenderOutput();
+        expect(result.type).toEqual('div');
+        expect(result.props.children).toEqual([]);
+    });
 });
-;
-
-var _temp = function () {
-  if (typeof __REACT_HOT_LOADER__ === 'undefined') {
-    return;
-  }
-}();
-
-;
