@@ -61,7 +61,6 @@ describe('utils', function () {
                 'description': 'Gimme yea name lad',
                 'type': 'string'
             },
-            'ngModelOptions': {},
             'key': ['name'],
             'type': 'text'
         }, {
@@ -71,7 +70,6 @@ describe('utils', function () {
                 'type': 'string',
                 'enum': ['undefined', 'null', 'NaN']
             },
-            'ngModelOptions': {},
             'key': ['gender'],
             'type': 'select',
             'titleMap': [{
@@ -91,7 +89,6 @@ describe('utils', function () {
                 'type': 'boolean',
                 'default': false
             },
-            'ngModelOptions': {},
             'key': ['overEighteen'],
             'type': 'checkbox'
         }, {
@@ -122,7 +119,6 @@ describe('utils', function () {
                     }
                 }
             },
-            'ngModelOptions': {},
             'type': 'fieldset',
             'items': [{
                 'title': 'Eye color',
@@ -131,7 +127,6 @@ describe('utils', function () {
                     'type': 'string',
                     'title': 'Eye color'
                 },
-                'ngModelOptions': {},
                 'key': ['attributes', 'eyecolor'],
                 'type': 'text'
             }, {
@@ -140,7 +135,6 @@ describe('utils', function () {
                     'type': 'string',
                     'title': 'Hair color'
                 },
-                'ngModelOptions': {},
                 'key': ['attributes', 'haircolor'],
                 'type': 'text'
             }, {
@@ -157,14 +151,12 @@ describe('utils', function () {
                         }
                     }
                 },
-                'ngModelOptions': {},
                 'type': 'fieldset',
                 'items': [{
                     'title': 'left',
                     'schema': {
                         'type': 'string'
                     },
-                    'ngModelOptions': {},
                     'key': ['attributes', 'shoulders', 'left'],
                     'type': 'text'
                 }, {
@@ -172,7 +164,6 @@ describe('utils', function () {
                     'schema': {
                         'type': 'string'
                     },
-                    'ngModelOptions': {},
                     'key': ['attributes', 'shoulders', 'right'],
                     'type': 'text'
                 }]
@@ -203,13 +194,12 @@ describe('utils', function () {
                 'description': 'Gimme yea name lad',
                 'type': 'string'
             },
-            'ngModelOptions': { 'updateOn': 'blur' },
             'foo': 'bar',
             'key': ['name'],
             'type': 'text'
         }];
 
-        var f = utils.getDefaults(schema, {}, { formDefaults: { foo: 'bar', ngModelOptions: { updateOn: 'blur' } } });
+        var f = utils.getDefaults(schema, {}, { formDefaults: { foo: 'bar' } });
         expect(f.form).toEqual(form);
     });
 
@@ -380,7 +370,7 @@ describe('utils', function () {
         expect(merged[0].items[0].items[0].readonly).toEqual(true);
     });
 
-    it('should select and set into objects and arrays', function () {
+    xit('should select and set into objects and arrays', function () {
         var schema = {
             'key': ['comments'],
             'add': 'New',
@@ -395,7 +385,6 @@ describe('utils', function () {
                     'title': 'Name',
                     'type': 'string'
                 },
-                'ngModelOptions': {},
                 'type': 'text'
             }, {
                 'key': ['comments', '', 'email'],
@@ -407,7 +396,6 @@ describe('utils', function () {
                     'pattern': '^\\S+@\\S+$',
                     'description': 'Email will be used for evil.'
                 },
-                'ngModelOptions': {},
                 'type': 'text'
             }, {
                 'key': ['comments', '', 'spam'],
@@ -418,8 +406,7 @@ describe('utils', function () {
                     'title': 'Spam',
                     'type': 'boolean',
                     'default': true
-                },
-                'ngModelOptions': {}
+                }
             }, {
                 'key': ['comments', '', 'comment'],
                 'type': 'textarea',
@@ -432,8 +419,7 @@ describe('utils', function () {
                     'type': 'string',
                     'maxLength': 20,
                     'validationMessage': 'Don\'t be greedy!'
-                },
-                'ngModelOptions': {}
+                }
             }],
             'title': 'comments',
             'required': true,
@@ -468,14 +454,12 @@ describe('utils', function () {
                     'required': ['name', 'comment']
                 }
             },
-            'ngModelOptions': {},
             'type': 'array'
         };
 
-        var model = {};
-        var list = utils.selectOrSet(schema, [{ key: 'sub', readonly: true }]);
-
-        //sub
+        //var list = utils.selectOrSet(schema, [{key: 'sub', readonly: true}]);
+        var merged = utils.merge(schema, [{ key: 'sub', readonly: true }]);
+        // sub
         expect(merged[0].readonly).toEqual(true);
 
         //array
@@ -485,12 +469,3 @@ describe('utils', function () {
         expect(merged[0].items[0].items[0].readonly).toEqual(true);
     });
 });
-;
-
-var _temp = function () {
-    if (typeof __REACT_HOT_LOADER__ === 'undefined') {
-        return;
-    }
-}();
-
-;
