@@ -1,8 +1,20 @@
 'use strict';
 
-var _lodash = require('lodash');
+var _isObject = require('lodash/isObject');
 
-var _lodash2 = _interopRequireDefault(_lodash);
+var _isObject2 = _interopRequireDefault(_isObject);
+
+var _cloneDeep = require('lodash/cloneDeep');
+
+var _cloneDeep2 = _interopRequireDefault(_cloneDeep);
+
+var _extend = require('lodash/extend');
+
+var _extend2 = _interopRequireDefault(_extend);
+
+var _isUndefined = require('lodash/isUndefined');
+
+var _isUndefined2 = _interopRequireDefault(_isUndefined);
 
 var _objectpath = require('objectpath');
 
@@ -21,7 +33,7 @@ function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { de
 //Evaluates an expression in a safe way
 function safeEval(condition, scope) {
     try {
-        var scope_safe = _lodash2.default.cloneDeep(scope);
+        var scope_safe = (0, _cloneDeep2.default)(scope);
         return (0, _notevil2.default)(condition, scope_safe);
     } catch (error) {
         return undefined;
@@ -61,7 +73,7 @@ var canonicalTitleMap = function canonicalTitleMap(titleMap, originalEnum) {
 //Creates a form object with all common properties
 var stdFormObj = function stdFormObj(name, schema, options) {
     options = options || {};
-    var f = options.global && options.global.formDefaults ? _lodash2.default.cloneDeep(options.global.formDefaults) : {};
+    var f = options.global && options.global.formDefaults ? (0, _cloneDeep2.default)(options.global.formDefaults) : {};
     if (options.global && options.global.supressPropertyTitles === true) {
         f.title = schema.title;
     } else {
@@ -100,11 +112,18 @@ var stdFormObj = function stdFormObj(name, schema, options) {
     }
     f.schema = schema;
 
-    // Ng model options doesn't play nice with undefined, might be defined
-    // globally though
-    f.ngModelOptions = f.ngModelOptions || {};
-
     return f;
+};
+
+var tBoolean = function tBoolean(name, schema, options) {
+    if (stripNullType(schema.type) === 'tBoolean' && !schema['enum']) {
+        var f = stdFormObj(name, schema, options);
+        f.key = options.path;
+        f.type = 'tBoolean';
+        options.lookup[_objectpath2.default.stringify(options.path)] = f;
+
+        return f;
+    }
 };
 
 var text = function text(name, schema, options) {
@@ -261,7 +280,8 @@ var defaults = {
     integer: [integer],
     boolean: [checkbox],
     array: [checkboxes, array],
-    date: [date]
+    date: [date],
+    tBoolean: [tBoolean]
 };
 
 function defaultFormDefinition(name, schema, options) {
@@ -278,8 +298,8 @@ function defaultFormDefinition(name, schema, options) {
             if (def) {
 
                 // Do we have form defaults in the schema under the x-schema-form-attribute?
-                if (def.schema['x-schema-form'] && _lodash2.default.isObject(def.schema['x-schema-form'])) {
-                    def = _lodash2.default.extend(def, def.schema['x-schema-form']);
+                if (def.schema['x-schema-form'] && (0, _isObject2.default)(def.schema['x-schema-form'])) {
+                    def = (0, _extend2.default)(def, def.schema['x-schema-form']);
                 }
                 return def;
             }
@@ -424,7 +444,6 @@ function merge(schema, form, ignore, options, readonly) {
         if (typeof obj === 'string') {
             obj = { key: obj };
         }
-        //console.log('obj', obj);
         if (obj.key) {
             if (typeof obj.key === 'string') {
                 obj.key = _objectpath2.default.parse(obj.key);
@@ -442,7 +461,7 @@ function merge(schema, form, ignore, options, readonly) {
             var str = _objectpath2.default.stringify(obj.key);
             var stdForm = lookup[str];
             stdForm.items.forEach(function (item) {
-                var o = _lodash2.default.cloneDeep(obj.itemForm);
+                var o = (0, _cloneDeep2.default)(obj.itemForm);
                 o.key = item.key;
                 obj.items.push(o);
             });
@@ -486,7 +505,7 @@ function merge(schema, form, ignore, options, readonly) {
 
         // Special case: checkbox
         // Since have to ternary state we need a default
-        if (obj.type === 'checkbox' && _lodash2.default.isUndefined(obj.schema['default'])) {
+        if (obj.type === 'checkbox' && (0, _isUndefined2.default)(obj.schema['default'])) {
             obj.schema['default'] = false;
         }
 
@@ -636,64 +655,3 @@ module.exports = {
     safeEval: safeEval,
     selectOrSet: selectOrSet
 };
-;
-
-var _temp = function () {
-    if (typeof __REACT_HOT_LOADER__ === 'undefined') {
-        return;
-    }
-
-    __REACT_HOT_LOADER__.register(safeEval, 'safeEval', 'src/utils.js');
-
-    __REACT_HOT_LOADER__.register(stripNullType, 'stripNullType', 'src/utils.js');
-
-    __REACT_HOT_LOADER__.register(enumToTitleMap, 'enumToTitleMap', 'src/utils.js');
-
-    __REACT_HOT_LOADER__.register(canonicalTitleMap, 'canonicalTitleMap', 'src/utils.js');
-
-    __REACT_HOT_LOADER__.register(stdFormObj, 'stdFormObj', 'src/utils.js');
-
-    __REACT_HOT_LOADER__.register(text, 'text', 'src/utils.js');
-
-    __REACT_HOT_LOADER__.register(number, 'number', 'src/utils.js');
-
-    __REACT_HOT_LOADER__.register(integer, 'integer', 'src/utils.js');
-
-    __REACT_HOT_LOADER__.register(date, 'date', 'src/utils.js');
-
-    __REACT_HOT_LOADER__.register(checkbox, 'checkbox', 'src/utils.js');
-
-    __REACT_HOT_LOADER__.register(select, 'select', 'src/utils.js');
-
-    __REACT_HOT_LOADER__.register(checkboxes, 'checkboxes', 'src/utils.js');
-
-    __REACT_HOT_LOADER__.register(fieldset, 'fieldset', 'src/utils.js');
-
-    __REACT_HOT_LOADER__.register(array, 'array', 'src/utils.js');
-
-    __REACT_HOT_LOADER__.register(defaults, 'defaults', 'src/utils.js');
-
-    __REACT_HOT_LOADER__.register(defaultFormDefinition, 'defaultFormDefinition', 'src/utils.js');
-
-    __REACT_HOT_LOADER__.register(getDefaults, 'getDefaults', 'src/utils.js');
-
-    __REACT_HOT_LOADER__.register(postProcessFn, 'postProcessFn', 'src/utils.js');
-
-    __REACT_HOT_LOADER__.register(appendRule, 'appendRule', 'src/utils.js');
-
-    __REACT_HOT_LOADER__.register(prependRule, 'prependRule', 'src/utils.js');
-
-    __REACT_HOT_LOADER__.register(traverseSchema, 'traverseSchema', 'src/utils.js');
-
-    __REACT_HOT_LOADER__.register(traverseForm, 'traverseForm', 'src/utils.js');
-
-    __REACT_HOT_LOADER__.register(merge, 'merge', 'src/utils.js');
-
-    __REACT_HOT_LOADER__.register(selectOrSet, 'selectOrSet', 'src/utils.js');
-
-    __REACT_HOT_LOADER__.register(validateBySchema, 'validateBySchema', 'src/utils.js');
-
-    __REACT_HOT_LOADER__.register(validate, 'validate', 'src/utils.js');
-}();
-
-;
