@@ -1,49 +1,48 @@
-/**
- * Created by steve on 20/09/15.
- */
-import React from 'react';
-import ComposedComponent from './ComposedComponent';
-import Checkbox from 'material-ui/Checkbox';
+import React, {Component} from 'react'
+import ComposedComponent from './ComposedComponent'
+import Checkbox from '@material-ui/core/Checkbox'
+import FormGroup from '@material-ui/core/FormGroup'
+import FormControlLabel from '@material-ui/core/FormControlLabel'
 
-const styles = {
-    error: {
-        color: 'rgb(244, 67, 54)',
-        fontSize: '12px',
-        lineHeight: '12px',
-    },
-    checkbox: {
-        marginTop: '14px',
-        marginBottom: '0px',
-        height: '72px',
-    },
-};
 
-class Checkbox2 extends React.Component {
-
-    constructor (props) {
-        super(props);
-        const {model, form} = this.props;
-        const {key} = form;
+class FormCheckbox extends Component {
+    constructor(props) {
+        super(props)
+        const {model, form} = this.props
+        const {key} = form
         //If a boolean is stored, use it; if not, if a boolean is defined as schema's default, use it.
-        const value = _.isBoolean(this.props.value)? this.props.value : (_.isBoolean(form.schema.default)? form.schema.default : undefined);
+        const value = typeof this.props.value === 'boolean'
+            ? this.props.value
+            : typeof form.schema.default === 'boolean'
+                ? form.schema.default
+                : undefined
         this.props.setDefault(key, model, form, value)
+    }
+
+    handleChange = e => {
+        this.props.onChangeValidate(e, 'yes')
+        // this.props.onChangeValidate(e, !this.props.value)
     }
 
     render() {
         return (
-            <div className={this.props.form.className} style={styles.checkbox}>
-                <Checkbox
-                    name={this.props.form.key.slice(-1)[0]}
-                    value={this.props.form.key.slice(-1)[0]}
-                    checked={this.props.value}
+            <FormGroup row>
+                <FormControlLabel
+                    className={this.props.form.className}
                     label={this.props.form.title}
-                    disabled={this.props.form.readonly}
-                    onCheck={(e, checked) => {this.props.onChangeValidate(e)}}
-                    />
-                <span style={styles.error}>{this.props.errorText}</span>
-             </div>
-        );
+                    control={
+                        <Checkbox
+                            name={this.props.form.key.slice(-1)[0]}
+                            value={this.props.form.key.slice(-1)[0]}
+                            checked={this.props.value || false}
+                            disabled={this.props.form.readonly}
+                            onChange={this.handleChange}
+                        />
+                    }
+                />
+            </FormGroup>
+        )
     }
 }
 
-export default ComposedComponent(Checkbox2);
+export default ComposedComponent(FormCheckbox)
