@@ -1,15 +1,11 @@
-/**
- * Created by steve on 20/09/15.
- */
 import React from 'react';
-import classNames from 'classnames';
+import Typography from '@material-ui/core/Typography';
 
 class Help extends React.Component{
     render() {
-        let classes = classNames(this.props.form.htmlClass);
-        //console.log('Help:', this.props);
+        let {description} = this.props.form
         return (
-            <div className={classes} dangerouslySetInnerHTML={{__html: this.props.form.description}}  />
+            <Typography variant="body2">{description}</Typography>
         )
     }
 }
