@@ -1,74 +1,67 @@
-/**
- * Created by steve on 15/09/15.
- */
-import React from 'react';
-import ComposedComponent from './ComposedComponent';
-import TextField from 'material-ui/TextField';
+import React, {Component} from 'react'
+import ComposedComponent from './ComposedComponent'
+import {TextField} from '@material-ui/core'
 
 /**
  * There is no default number picker as part of Material-UI.
  * Instead, use a TextField and validate.
  */
-class Number extends React.Component {
+class Number extends Component {
 
     constructor(props) {
-        super(props);
-        this.preValidationCheck = this.preValidationCheck.bind(this);
+        super(props)
         this.state = {
-            lastSuccessfulValue : this.props.value
+            lastSuccessfulValue: this.props.value
+        }
+        this.numberField = React.createRef()
+    }
+
+    static getDerivedStateFromProps(nextProps) {
+        return {
+            lastSuccessfulValue: nextProps.value
         }
     }
 
-    componentWillReceiveProps(nextProps) {
-      this.setState({
-        lastSuccessfulValue: nextProps.value
-      });
-    }
+    isNumeric = n => !isNaN(parseFloat(n)) && isFinite(n)
 
-    isNumeric(n) {
-        return !isNaN(parseFloat(n)) && isFinite(n);
-    }
-
-    isEmpty(n) {
-        return (!n || 0 === n.length);
-    }
+    isEmpty = n => (!n || 0 === n.length)
 
     /**
      * Prevent the field from accepting non-numeric characters.
      * @param e
      */
-    preValidationCheck(e) {
+    preValidationCheck = (e) => {
         if (this.isNumeric(e.target.value)) {
             this.setState({
                 lastSuccessfulValue: e.target.value
-            });
-            this.props.onChangeValidate(e);
+            })
+            this.props.onChangeValidate(e)
         } else if (this.isEmpty(e.target.value)) {
             this.setState({
                 lastSuccessfulValue: e.target.value
-            });
-            this.props.onChangeValidate(e);
+            })
+            this.props.onChangeValidate(e)
         } else {
-            this.refs.numberField.value = this.state.lastSuccessfulValue;
+            this.numberField.current.value = this.state.lastSuccessfulValue
         }
     }
 
     render() {
+        let {form, error} = this.props
         return (
-            <div className={this.props.form.htmlClass}>
-                <TextField
-                    type={this.props.form.type}
-                    floatingLabelText={this.props.form.title}
-                    hintText={this.props.form.placeholder}
-                    errorText={this.props.error || this.props.errorText}
-                    onChange={this.preValidationCheck}
-                    value={this.state.lastSuccessfulValue}
-                    ref="numberField"
-                    disabled={this.props.form.readonly}
-                    style={this.props.form.style || {width: '100%'}}/>
-            </div>
-        );
+            <TextField
+                type={form.type}
+                label={form.title}
+                placeholder={form.placeholder}
+                helperText={error || form.description}
+                error={!!error}
+                onChange={this.preValidationCheck}
+                value={this.state.lastSuccessfulValue}
+                ref={this.numberField}
+                disabled={form.readonly}
+                fullWidth/>
+        )
     }
 }
 
-export default ComposedComponent(Number);
+export default ComposedComponent(Number)
