@@ -1,41 +1,35 @@
-/**
- * Created by steve on 15/09/15.
- */
-import React from 'react';
-import ComposedComponent from './ComposedComponent';
-import MenuItem from 'material-ui/MenuItem';
-import SelectField from 'material-ui/SelectField';
-import _ from 'lodash';
+import React, {Component} from 'react'
+import ComposedComponent from './ComposedComponent'
+import {MuiSelect, InputLabel, MenuItem, FormControl} from '@material-ui/core'
+import {isEmpty} from 'lodash'
 
-class Select extends React.Component {
 
+class Select extends Component {
     constructor(props) {
-        super(props);
-        this.onSelected = this.onSelected.bind(this);
+        super(props)
 
-        const {model, form} = this.props;
-        const {key} = form;
+        const {model, form} = this.props
+        const {key} = form
 
-        const storedValue = model && this.getModelKey(model, key) || false;
-        const defaultValue = form.schema.default || false;
-        const value = !(_.isEmpty(storedValue)) && storedValue || defaultValue;
+        const storedValue = model && Select.getModelKey(model, key) || false
+        const defaultValue = form.schema.default || false
+        const value = !(isEmpty(storedValue)) && storedValue || defaultValue
 
         this.props.setDefault(key, model, form, value)
         this.state = {
             currentValue: value,
-        };
-    }
-
-    componentWillReceiveProps(nextProps) {
-        if (nextProps.model && nextProps.form.key) {
-            this.setState({
-                currentValue: this.getModelKey(nextProps.model, nextProps.form.key)
-                || (nextProps.form.titleMap != null ? nextProps.form.titleMap[0].value : '')
-            });
         }
     }
 
-    getModelKey(model, key) {
+    static getDerivedStateFromProps(props) {
+        if (props.model && props.form.key) {
+            return {
+                currentValue: Select.getModelValue(props.model, props.form)
+            }
+        }
+    }
+
+    static getModelKey = (model, key) => {
         if (Array.isArray(key)) {
             return key.reduce((cur, nxt) => (cur[nxt] || {}), model);
         } else {
@@ -43,40 +37,32 @@ class Select extends React.Component {
         }
     }
 
-    onSelected(event, selectedIndex, menuItem) {
+    static getModelValue = (model, {key, titleMap}) =>
+        Select.getModelKey(model, key) || (titleMap != null ? titleMap[0].value : '')
 
-        this.setState({
-            currentValue: menuItem
-        });
-        event.target.value = menuItem;
-        this.props.onChangeValidate(event);
+    onSelected = ({target: {value: currentValue}}) => {
+        this.setState({currentValue})
+        this.props.onChangeValidate(event)
     }
 
     render() {
-        const menuItems = this.props.form.titleMap.map((item, idx) => (
-            <MenuItem key={idx}
-                      primaryText={item.name}
-                      value={item.value} />
-        ));
-
+        const {form} = this.props
+        const menuItems = form.titleMap.map((item, idx) => (
+            <MenuItem key={idx} value={item.value}>{item.name}</MenuItem>
+        ))
         return (
-            <div className={this.props.form.htmlClass}>
-                <SelectField
-                    value={this.state.currentValue}
-                    floatingLabelText={this.props.form.title}
-                    disabled={this.props.form.readonly}
-                    onChange={this.onSelected}
-                    fullWidth >
-
+            <FormControl style={{width: '100%'}}>
+                <InputLabel>{form.title}</InputLabel>
+                <MuiSelect
+                    value={this.state.currentValue || ''}
+                    placeholder={form.title}
+                    disabled={form.readonly}
+                    onChange={this.onSelected}>
                     {menuItems}
-                </SelectField>
-            </div>
-        );
+                </MuiSelect>
+            </FormControl>
+        )
     }
 }
 
-// Select.propTypes = {
-//
-// };
-
-export default ComposedComponent(Select);
+export default ComposedComponent(Select)
