@@ -1,42 +1,43 @@
+import React from 'react';
+import ShallowRenderer from 'react-test-renderer/shallow';
+import ComposedComponent from '../ComposedComponent';
+import Text from '../Text';
+
 jest.dontMock('../ComposedComponent');
 jest.dontMock('../utils');
 jest.dontMock('lodash');
 
-var React = require('react');
-var TestUtils = require('react-addons-test-utils');
-import ComposedComponent from '../ComposedComponent';
-import Text from '../Text';
 
-describe('ComposedComponent', function() {
+describe('ComposedComponent', function () {
 
-    it('shows default value at text field', function() {
-        var shallowRenderer = TestUtils.createRenderer();
-        var cfg = {
-            form: {
-                key: ['name'],
-                schema: {
-                    default: 'steeve',
-                    title: 'name',
-                    type: 'String',
-                },
-                type: 'text',
-                title: 'name',
-            },
-            model: {name: 'steeve'},
-            mapper: {}
-          };
+  it('shows default value at text field', function () {
+    const renderer = new ShallowRenderer();
+    let cfg = {
+      form: {
+        key: ['name'],
+        schema: {
+          default: 'steeve',
+          title: 'name',
+          type: 'String',
+        },
+        type: 'text',
+        title: 'name',
+      },
+      model: {name: 'steeve'},
+      mapper: {}
+    };
 
-        var Composed = ComposedComponent(Text);
+    let Composed = ComposedComponent(Text);
 
-        shallowRenderer.render( 
-            <Composed 
-                form={cfg.form} 
-                model={cfg.model}
-                mapper={cfg.mapper}
-        />);
+    renderer.render(
+      <Composed
+        form={cfg.form}
+        model={cfg.model}
+        mapper={cfg.mapper}
+      />);
 
-        var result = shallowRenderer.getRenderOutput();
+    let result = renderer.getRenderOutput();
 
-        expect(result.props.value).toEqual('steeve');
-    });
+    expect(result.props.value).toEqual('steeve');
+  });
 });
