@@ -1,4 +1,7 @@
-import _ from 'lodash';
+import isObject from 'lodash/isObject';
+import cloneDeep from 'lodash/cloneDeep'
+import extend from 'lodash/extend'
+import isUndefined from 'lodash/isUndefined'
 import ObjectPath from 'objectpath';
 import tv4 from 'tv4';
 import notevil from 'notevil';
@@ -6,7 +9,7 @@ import notevil from 'notevil';
 //Evaluates an expression in a safe way
 function safeEval(condition, scope) {
     try {
-        const scope_safe = _.cloneDeep(scope);
+        const scope_safe = cloneDeep(scope);
         return notevil(condition, scope_safe);
     } catch (error) {
         return undefined
@@ -49,7 +52,7 @@ var canonicalTitleMap = function(titleMap, originalEnum) {
 //Creates a form object with all common properties
 var stdFormObj = function(name, schema, options) {
     options = options || {};
-    var f = options.global && options.global.formDefaults ? _.cloneDeep(options.global.formDefaults) : {};
+    var f = options.global && options.global.formDefaults ? cloneDeep(options.global.formDefaults) : {};
     if (options.global && options.global.supressPropertyTitles === true) {
         f.title = schema.title;
     } else {
@@ -70,11 +73,18 @@ var stdFormObj = function(name, schema, options) {
     if (schema.enumNames) { f.titleMap = canonicalTitleMap(schema.enumNames, schema['enum']); }
     f.schema = schema;
 
-    // Ng model options doesn't play nice with undefined, might be defined
-    // globally though
-    f.ngModelOptions = f.ngModelOptions || {};
-
     return f;
+};
+
+var tBoolean = function(name, schema, options) {
+    if (stripNullType(schema.type) === 'tBoolean' && !schema['enum']) {
+        var f = stdFormObj(name, schema, options);
+        f.key  = options.path;
+        f.type = 'tBoolean';
+        options.lookup[ObjectPath.stringify(options.path)] = f;
+        
+        return f;
+    }
 };
 
 var text = function(name, schema, options) {
@@ -232,7 +242,8 @@ var defaults = {
     integer: [integer],
     boolean: [checkbox],
     array:   [checkboxes, array],
-    date:    [date]
+    date:    [date],
+    tBoolean:[tBoolean]
 };
 
 function defaultFormDefinition(name, schema, options) {
@@ -249,8 +260,8 @@ function defaultFormDefinition(name, schema, options) {
             if (def) {
 
                 // Do we have form defaults in the schema under the x-schema-form-attribute?
-                if (def.schema['x-schema-form'] && _.isObject(def.schema['x-schema-form'])) {
-                    def = _.extend(def, def.schema['x-schema-form']);
+                if (def.schema['x-schema-form'] && isObject(def.schema['x-schema-form'])) {
+                    def = extend(def, def.schema['x-schema-form']);
                 }
                 return def;
             }
@@ -395,7 +406,6 @@ function merge(schema, form, ignore, options, readonly) {
         if (typeof obj === 'string') {
             obj = {key: obj};
         }
-        //console.log('obj', obj);
         if (obj.key) {
             if (typeof obj.key === 'string') {
                 obj.key = ObjectPath.parse(obj.key);
@@ -413,7 +423,7 @@ function merge(schema, form, ignore, options, readonly) {
             var str = ObjectPath.stringify(obj.key);
             var stdForm = lookup[str];
             stdForm.items.forEach(function(item) {
-                var o = _.cloneDeep(obj.itemForm);
+                var o = cloneDeep(obj.itemForm);
                 o.key = item.key;
                 obj.items.push(o);
             });
@@ -458,7 +468,7 @@ function merge(schema, form, ignore, options, readonly) {
 
         // Special case: checkbox
         // Since have to ternary state we need a default
-        if (obj.type === 'checkbox' && _.isUndefined(obj.schema['default'])) {
+        if (obj.type === 'checkbox' && isUndefined(obj.schema['default'])) {
             obj.schema['default'] = false;
         }
 
